+++ py4macro/py4macro.py
@@ -151,24 +151,24 @@ ppp_definitions="""
     | `country`: 国名
     | `iso`: ISO国コード
     | `currency_code`: 通貨コード
-    | `bm_local_price`: ビッグマックの自国通貨価格
+    | `bm_lcu`: ビッグマックの価格（自国通貨単位，Local Currency Unit）
     | `exr`: 名目為替レート（自国通貨単位/米ドル）
-    | `ppp`: 購買力平価に基づく交換率（自国通貨単位/international dollar）
-    | `rgdppc`: 実質一人当たりGDP（購買力平価, international dollar）
-    | `ngdppc`: 名目一人当たりGDP（購買力平価, international dollar）
-    | `rgdppc_lcu`: 実質一人当たりGDP（Local Currency Unit, 自国通貨単位）
-    | `ngdppc_lcu`: 名目一人当たりGDP（Local Currency Univ, 自国通貨単位）
+    | `ppp`: 購買力平価レート（自国通貨単位/international dollar）
+    | `rgdppc`: 実質一人当たりGDP（購買力平価レート, international dollar）
+    | `ngdppc`: 名目一人当たりGDP（購買力平価レート, international dollar）
+    | `rgdppc_lcu`: 実質一人当たりGDP（自国通貨単位，Local Currency Unit）
+    | `ngdppc_lcu`: 名目一人当たりGDP（自国通貨単位，Local Currency Unit）
     |
     | * 年次データ
     |
     | ＜出典＞
-    | `bm_local_price`: https://github.com/TheEconomist/big-mac-data
-    | `exr`: `PA.NUS.FCRF`, the World Bank
-    | `ppp`: `PA.NUS.PPP`,  the World Bank
-    | `rgdppc: `rgdppc_local`/`ppp`
-    | `ngdppc`: `ngdppc_local`/`ppp`
-    | `rgdppc_local: `NY.GDP.PCAP.KN`, the World Bank
-    | `ngdppc_local`: `NY.GDP.PCAP.CN`, the World Bank"""
+    | `bm_lcu`: https://github.com/TheEconomist/big-mac-data
+    | `exr`: データID `PA.NUS.FCRF`, the World Bank
+    | `ppp`: データID `PA.NUS.PPP`,  the World Bank
+    | `rgdppc: `rgdppc_local`/`ppp`として計算
+    | `ngdppc`: `ngdppc_local`/`ppp`として計算
+    | `rgdppc_local: データID `NY.GDP.PCAP.KN`, the World Bank
+    | `ngdppc_local`: データID `NY.GDP.PCAP.CN`, the World Bank"""
 
 # ===== Helper functions =======================================================================